Moved AgentProcessRepository to core package
See gh-1024
