+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/api/common/ranking/Ranker.kt
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.embabel.agent.api.common.ranking
 
 import com.embabel.common.core.types.*
