+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/core/AgentProcessRepository.kt
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.embabel.agent.spi
-
-import com.embabel.agent.core.AgentProcess
+package com.embabel.agent.core
 
 /**
  * Repository for agent processes.
