+++ embabel-agent-api/src/main/kotlin/com/embabel/agent/spi/support/SpringContextPlatformServices.kt
@@ -21,8 +21,8 @@ import com.embabel.agent.api.common.PlatformServices
 import com.embabel.agent.api.common.autonomy.Autonomy
 import com.embabel.agent.api.event.AgenticEventListener
 import com.embabel.agent.core.AgentPlatform
+import com.embabel.agent.core.AgentProcessRepository
 import com.embabel.agent.core.expression.LogicalExpressionParser
-import com.embabel.agent.spi.AgentProcessRepository
 import com.embabel.agent.spi.LlmOperations
 import com.embabel.agent.spi.OperationScheduler
 import com.embabel.agent.spi.expression.spel.SpelLogicalExpressionParser
